requirements.txt
3==0.0.0
blinker==1.9.0
boto==2.49.0
boto3==1.35.57
botocore==1.35.57
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fakeredis==2.26.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
googlemaps==4.10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
polyline==2.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
